Version 2.0
Updated with errors fixed and few functionalities added. (previous code was not fulfilling few requirements such as considering longest prefix for minimum price) 
Provision for CSV Excel file file to be added where file path needs to be pasted is mentioned as a comment but can be easily converted to statement.