Phone Number Routing
Different version of TeleRout.py, a code with a different form of solution for implementing the task presented. 
Has different way of presenting the output, with unit tests working as well. 
Has provision of importing CSV excel file, which gives a provision for huge number of Operators and Price lists to be created.